Refactor of plugin package
Move all rpc related code to another file and remove passing errors
messages via rpc arguments but instead return an error on the rpc server
methods.
